refactor: make image imports in TS examples work with Astro
