+++ frontend/types.d.ts
@@ -1,9 +1,9 @@
-declare module '*.png' {
+declare module '*.png?url' {
   const value: string;
   export = value;
 }
 
-declare module '*.jpg' {
+declare module '*.jpg?url' {
   const value: string;
   export = value;
 }
